+++ updates/0004/psycopg2-dependencies.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/bash
+set -e
+
+sudo apt install python3-dev
+sudo apt install libpq-dev